Consolidates isOutOfBounds
